Put team creation back
