+++ forge/ee/routes/sso/auth.js
@@ -4,6 +4,8 @@ const { Authenticator } = require('@fastify/passport')
 const { MultiSamlStrategy } = require('@node-saml/passport-saml')
 const fp = require('fastify-plugin')
 
+const createTeamForUser = require('../../../lib/userTeam')
+
 const generatePassword = () => {
     const charList = '0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz~!@-#$'
     return Array.from(crypto.randomFillSync(new Uint32Array(8))).map(x => charList[x % charList.length]).join('')
@@ -131,6 +133,9 @@ module.exports = fp(async function (app, opts) {
                                 done(err)
                                 return
                             }
+                        } else {
+                            // no SSO Group mapping so create team
+                            await createTeamForUser(app, newUser)
                         }
                         request.session.newSSOUser = true
                         done(null, newUser)
